Ditched the circle, made progress
I've decided to ditch the circle theme for my site, as CSS animation of
psudo-elements is not supported on many browsers. I've added the new
pages for the site, as well as fleshed out the style.css file. I make
reference to .png files that have not bee created yet, so you're not
missing anything if you can't get them to load.